refactor(worktree): decouple file setup from docker plugin
copy_files, symlink_files, and symlink_dirs were running only via the
docker plugin's OnPostCreate hook. If the docker plugin was disabled
(config enabled = false), uninstalled, or failed to initialize, the
hook never fired and credential files / symlinks never landed in the
worktree. Users hit this most often with --no-docker, leaving Rails
unable to boot in the new worktree.

Move the file setup logic into a new internal/worktree.SetupFiles
function and call it unconditionally from setupCreatedWorktree. The
docker plugin's OnPostCreate methods become no-ops for file setup
(externalStrategy still persists the env var and emits the env
directive — those are legitimate docker concerns).

Config still lives at [plugins.docker.external] — no migration needed
for existing users. The decoupling happens in the call graph, not the
config schema.

Closes #27.
