+++ plugins/docker/agent_external_test.go
@@ -171,58 +171,6 @@ func TestAgentExternalStrategy_OnPostCreate_NoPaths(t *testing.T) {
 	}
 }
 
-func TestAgentExternalStrategy_OnPostCreate_CopiesFiles(t *testing.T) {
-	tmpDir := t.TempDir()
-
-	mainPath := filepath.Join(tmpDir, "main")
-	_ = os.MkdirAll(filepath.Join(mainPath, "config"), 0755)
-	_ = os.WriteFile(filepath.Join(mainPath, "config", "secret.key"), []byte("secret"), 0600)
-
-	newPath := filepath.Join(tmpDir, "worktree")
-	_ = os.MkdirAll(newPath, 0755)
-
-	enabled := true
-	cfg := &config.Config{
-		ProjectName: "myapp",
-		Plugins: config.PluginsConfig{
-			Docker: config.DockerPluginConfig{
-				Mode: "external",
-				External: &config.ExternalComposeConfig{
-					Path:      tmpDir,
-					EnvVar:    "APP_DIR",
-					Services:  []string{"app"},
-					CopyFiles: []string{"config/secret.key"},
-					Agent: &config.AgentStackConfig{
-						Enabled:      &enabled,
-						MaxSlots:     3,
-						Services:     []string{"app"},
-						TemplatePath: "agent-stacks/template.yml",
-					},
-				},
-			},
-		},
-	}
-
-	s := newAgentExternalStrategy(cfg)
-	ctx := &hooks.Context{
-		Worktree:     "worktree",
-		WorktreePath: newPath,
-		MainPath:     mainPath,
-	}
-
-	if err := s.OnPostCreate(ctx); err != nil {
-		t.Fatalf("OnPostCreate() error = %v", err)
-	}
-
-	data, err := os.ReadFile(filepath.Join(newPath, "config", "secret.key"))
-	if err != nil {
-		t.Fatalf("Failed to read copied file: %v", err)
-	}
-	if string(data) != "secret" {
-		t.Errorf("Expected 'secret', got %q", string(data))
-	}
-}
-
 func TestAgentExternalStrategy_ComposeProjectName(t *testing.T) {
 	cfg := newTestAgentConfig(t)
 	s := newAgentExternalStrategy(cfg)
@@ -359,65 +307,6 @@ func TestResolveComposePath(t *testing.T) {
 				t.Errorf("resolveComposePath(%q) = %q, want %q", tt.path, got, tt.path)
 			}
 		})
-	}
-}
-
-func TestSetupWorktreeFiles(t *testing.T) {
-	tmpDir := t.TempDir()
-
-	mainPath := filepath.Join(tmpDir, "main")
-	_ = os.MkdirAll(filepath.Join(mainPath, "config"), 0755)
-	_ = os.WriteFile(filepath.Join(mainPath, "config", "secret.key"), []byte("secret"), 0600)
-	_ = os.MkdirAll(filepath.Join(mainPath, "vendor", "bundle"), 0755)
-
-	newPath := filepath.Join(tmpDir, "worktree")
-	_ = os.MkdirAll(newPath, 0755)
-
-	ext := &config.ExternalComposeConfig{
-		CopyFiles:   []string{"config/secret.key"},
-		SymlinkDirs: []string{"vendor/bundle"},
-	}
-
-	err := setupWorktreeFiles(ext, newPath, mainPath)
-	if err != nil {
-		t.Fatalf("setupWorktreeFiles() error = %v", err)
-	}
-
-	// Verify copied file
-	data, err := os.ReadFile(filepath.Join(newPath, "config", "secret.key"))
-	if err != nil {
-		t.Fatalf("Failed to read copied file: %v", err)
-	}
-	if string(data) != "secret" {
-		t.Errorf("Expected 'secret', got %q", string(data))
-	}
-
-	// Verify symlink
-	info, err := os.Lstat(filepath.Join(newPath, "vendor", "bundle"))
-	if err != nil {
-		t.Fatalf("Failed to stat symlink: %v", err)
-	}
-	if info.Mode()&os.ModeSymlink == 0 {
-		t.Error("Expected vendor/bundle to be a symlink")
-	}
-}
-
-func TestSetupWorktreeFiles_MissingSource(t *testing.T) {
-	tmpDir := t.TempDir()
-
-	mainPath := filepath.Join(tmpDir, "main")
-	_ = os.MkdirAll(mainPath, 0755)
-
-	newPath := filepath.Join(tmpDir, "worktree")
-	_ = os.MkdirAll(newPath, 0755)
-
-	ext := &config.ExternalComposeConfig{
-		CopyFiles: []string{"nonexistent.key"},
-	}
-
-	err := setupWorktreeFiles(ext, newPath, mainPath)
-	if err == nil {
-		t.Error("Expected error for missing source file, got nil")
 	}
 }
 
@@ -737,118 +626,6 @@ func TestHasActiveAgentSlot_NoAgentConfig(t *testing.T) {
 
 	if HasActiveAgentSlot(cfg, "/tmp/wt") {
 		t.Error("HasActiveAgentSlot should be false without agent config")
-	}
-}
-
-func TestSetupWorktreeFiles_SymlinkFiles(t *testing.T) {
-	tmpDir := t.TempDir()
-
-	mainPath := filepath.Join(tmpDir, "main")
-	_ = os.MkdirAll(filepath.Join(mainPath, "config", "credentials"), 0755)
-	_ = os.WriteFile(filepath.Join(mainPath, "config", "credentials", "dev.key"), []byte("devkey"), 0600)
-
-	newPath := filepath.Join(tmpDir, "worktree")
-	_ = os.MkdirAll(newPath, 0755)
-
-	ext := &config.ExternalComposeConfig{
-		SymlinkFiles: []string{"config/credentials/dev.key"},
-	}
-
-	err := setupWorktreeFiles(ext, newPath, mainPath)
-	if err != nil {
-		t.Fatalf("setupWorktreeFiles() error = %v", err)
-	}
-
-	// Verify it's a symlink, not a copy
-	dst := filepath.Join(newPath, "config", "credentials", "dev.key")
-	info, err := os.Lstat(dst)
-	if err != nil {
-		t.Fatalf("Failed to stat symlinked file: %v", err)
-	}
-	if info.Mode()&os.ModeSymlink == 0 {
-		t.Error("Expected config/credentials/dev.key to be a symlink")
-	}
-
-	// Verify content is accessible through the symlink
-	data, err := os.ReadFile(dst)
-	if err != nil {
-		t.Fatalf("Failed to read through symlink: %v", err)
-	}
-	if string(data) != "devkey" {
-		t.Errorf("Expected 'devkey', got %q", string(data))
-	}
-}
-
-func TestSetupWorktreeFiles_AllThreeTypes(t *testing.T) {
-	tmpDir := t.TempDir()
-
-	mainPath := filepath.Join(tmpDir, "main")
-	_ = os.MkdirAll(filepath.Join(mainPath, "config"), 0755)
-	_ = os.WriteFile(filepath.Join(mainPath, "config", "settings.yml"), []byte("settings"), 0644)
-	_ = os.WriteFile(filepath.Join(mainPath, "config", "secret.key"), []byte("secret"), 0600)
-	_ = os.MkdirAll(filepath.Join(mainPath, "vendor", "bundle"), 0755)
-
-	newPath := filepath.Join(tmpDir, "worktree")
-	_ = os.MkdirAll(newPath, 0755)
-
-	ext := &config.ExternalComposeConfig{
-		CopyFiles:    []string{"config/settings.yml"},
-		SymlinkFiles: []string{"config/secret.key"},
-		SymlinkDirs:  []string{"vendor/bundle"},
-	}
-
-	err := setupWorktreeFiles(ext, newPath, mainPath)
-	if err != nil {
-		t.Fatalf("setupWorktreeFiles() error = %v", err)
-	}
-
-	// Verify copied file is a regular file
-	copyInfo, err := os.Lstat(filepath.Join(newPath, "config", "settings.yml"))
-	if err != nil {
-		t.Fatalf("Failed to stat copied file: %v", err)
-	}
-	if copyInfo.Mode()&os.ModeSymlink != 0 {
-		t.Error("Expected config/settings.yml to be a regular file (copy), not a symlink")
-	}
-
-	// Verify symlinked file
-	fileInfo, err := os.Lstat(filepath.Join(newPath, "config", "secret.key"))
-	if err != nil {
-		t.Fatalf("Failed to stat symlinked file: %v", err)
-	}
-	if fileInfo.Mode()&os.ModeSymlink == 0 {
-		t.Error("Expected config/secret.key to be a symlink")
-	}
-
-	// Verify symlinked directory
-	dirInfo, err := os.Lstat(filepath.Join(newPath, "vendor", "bundle"))
-	if err != nil {
-		t.Fatalf("Failed to stat symlinked dir: %v", err)
-	}
-	if dirInfo.Mode()&os.ModeSymlink == 0 {
-		t.Error("Expected vendor/bundle to be a symlink")
-	}
-}
-
-func TestSetupWorktreeFiles_SymlinkFiles_MissingSource(t *testing.T) {
-	tmpDir := t.TempDir()
-
-	mainPath := filepath.Join(tmpDir, "main")
-	_ = os.MkdirAll(mainPath, 0755)
-
-	newPath := filepath.Join(tmpDir, "worktree")
-	_ = os.MkdirAll(newPath, 0755)
-
-	ext := &config.ExternalComposeConfig{
-		SymlinkFiles: []string{"nonexistent.key"},
-	}
-
-	err := setupWorktreeFiles(ext, newPath, mainPath)
-	if err == nil {
-		t.Error("Expected error for missing source file, got nil")
-	}
-	if !strings.Contains(err.Error(), "source not found") {
-		t.Errorf("Expected 'source not found' in error, got %q", err.Error())
 	}
 }
