+++ plugins/docker/external.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/lost-in-the/grove/internal/cli"
 	"github.com/lost-in-the/grove/internal/config"
-	"github.com/lost-in-the/grove/internal/fsutil"
 	"github.com/lost-in-the/grove/internal/hooks"
 	"github.com/lost-in-the/grove/internal/worktree"
 )
@@ -91,7 +90,7 @@ func (s *externalStrategy) OnPostCreate(ctx *hooks.Context) error {
 	}
 	s.emitEnvDirective(ctx.WorktreePath)
 
-	return setupWorktreeFiles(s.ext, ctx.WorktreePath, ctx.MainPath)
+	return nil
 }
 
 func (s *externalStrategy) Up(worktreePath string, detach bool) error {
@@ -349,34 +348,4 @@ func (s *externalStrategy) getAutoStop() bool {
 	}
 	// External mode defaults to true for auto_stop (unlike local's false)
 	return true
-}
-
-// copyFile copies a single file from src to dst, creating parent directories as needed.
-func copyFile(src, dst string) error {
-	return fsutil.CopyFile(src, dst)
-}
-
-// createSymlink creates a symbolic link from src to dst, creating parent directories
-// as needed. If dst already exists and is a symlink, it is replaced.
-func createSymlink(src, dst string) error {
-	if _, err := os.Lstat(src); err != nil {
-		return fmt.Errorf("source not found: %w", err)
-	}
-
-	if err := os.MkdirAll(filepath.Dir(dst), 0o755); err != nil {
-		return fmt.Errorf("failed to create parent directory: %w", err)
-	}
-
-	// Remove existing symlink if present
-	if info, err := os.Lstat(dst); err == nil {
-		if info.Mode()&os.ModeSymlink != 0 {
-			if err := os.Remove(dst); err != nil {
-				return fmt.Errorf("failed to remove existing symlink %s: %w", dst, err)
-			}
-		} else {
-			return fmt.Errorf("%s already exists and is not a symlink", dst)
-		}
-	}
-
-	return os.Symlink(src, dst)
 }
